[ci] add clang_tidy.sh helper and fix clang-tidy findings
- scripts/clang_tidy.sh: drive clang-tidy-19 from compile_commands.json so
  only files actually built are linted; exclude third_party and vendored
  embedder sources. Lint workflow now calls the script.
- shell/app.cc: cast duration to double before frame-time subtraction.
- logging_handler: make HandleMethodCall static (no instance state used).
- platform_view_touch.cc: add missing logging/logging.h include.
- wayland/display: pass strings by const-ref; use emplace_back for
  pending_app_list; rename display_handle_scale param to match
  wl_output_listener signature (int32_t factor).

+++ shell/app.cc
@@ -92,7 +92,8 @@ int App::Loop() const {
   const auto elapsed = end_time - start_time;
 
   const auto frame_time = 1000.0 / m_wayland_display->GetMaxRefreshRate();
-  if (const auto sleep_time = frame_time - elapsed; sleep_time > 0) {
+  if (const auto sleep_time = frame_time - static_cast<double>(elapsed);
+      sleep_time > 0) {
 #if BUILD_WATCHDOG
     m_watch_dog->pet();
 #endif
